mdt_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
django-extensions==4.0
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
drf-yasg==1.21.10
h11==0.14.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==24.2
prompt_toolkit==3.0.50
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
selenium==4.31.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
